use pytest fixtures
